chore: revert ipairs update
may be necessary in the future but lua already reliably uses ipairs order for array-like tables, and some objects tables actually aren't array-like so the change may not always be safe. the only safe ipairs-based solution would be to iterate over the entire ipairs, then check if its size was equal to the table size, and if not return pairs. but that's dumb and again, not necessary currently, though may become necessary should lua's promise that the pairs sort order is undetermined ever become true for arrays